refactor: Extract HarvestLoader into its own file
